requirements.txt
attrs==23.1.0
certifi==2023.11.17
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
outcome==1.3.0.post0
pip==21.1.2
PySocks==1.7.1
robotframework-pageobjectlibrary==1.1
robotframework-pythonlibcore==4.3.0
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==6.2.0
robotframework==6.1.1
selenium==4.16.0
setuptools==57.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio-websocket==0.11.1
trio==0.23.2
urllib3==2.1.0
wheel==0.36.2
wsproto==1.2.0
